Improve the build process.
Tests can be automatically build now, and header dependencies
of the SUT(System Under Test) are considered.
Split the source code into utils/ and src/ (core logic).
Everything in utils/ is linked with the tests for convenience.
Regarding the new Makefiles there are still things to be desired.
There are some hardcoded things in them.
CFLAGS is shared between the main, src/, utils/ Makefiles.
src's makefile is hardlinked in utils/.